usb_cam: refactor to support image dumps and ros msg transfter
Current version of the code had removed support for transferring
captured images are ros sensor msgs and instead used the custom
DataInFile msg to dump imgs to disk.

This patch refactors the code to support both methods, dumping imgs to
disk and also transfer as std ros sensor msgs.

The method to use is defined by setting the dump_to_disk param. If true,
the DataInFile is used to dump imgs to disk. If false, the std ros
sensor img msg is used to transfer imgs across ros topics.

Signed-off-by: Nitheesh K L <nitheesh.k.l@intel.com>