visualizer: add web visualizer
This patch add the python flask based web service to visualize a summary
of all the raw driving data collected.

This patch is add subset of contents and features from a larger
visualizer dashbaord. Therefore a couple of web links may be broken
until the full functionality of the web dashboard is integrated.

Signed-off-by: Nitheesh K L <nitheesh.k.l@intel.com>